Extract chart data and behavior into separate class
